API処理修正
